Add maven setup for project
