+++ src/main/java/dev/jbang/fmt/CodeRange.java
