+++ src/main/java/dev/jbang/fmt/KeyValueConsumer.java
