+++ src/main/resources/eclipse.xml
