+++ src/main/resources/jbang.xml
